feat: 메인화면, 홈화면, 알림화면 딥링크 추가
