+++ app/build.gradle
@@ -21,6 +21,10 @@ android {
         versionCode 1
         versionName "1.0.0"
 
+        manifestPlaceholders = [
+                schemePrefix          : schemes['bestfriends']
+        ]
+
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
@@ -48,6 +52,13 @@ android {
     }
     buildFeatures {
         dataBinding true
+    }
+}
+
+kapt {
+    arguments {
+        arg("deepLink.incremental", "true")
+        arg("deepLink.customAnnotations", "com.yapp.android2.deeplink.DeepLinkPrefixSpec")
     }
 }
