+++ app/src/main/AndroidManifest.xml
@@ -41,6 +41,19 @@
         </activity>
 
         <activity
+            android:name=".deeplink.DeepLinkActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW" />
+
+                <category android:name="android.intent.category.DEFAULT" />
+                <category android:name="android.intent.category.BROWSABLE" />
+
+                <data android:scheme="${schemePrefix}" />
+            </intent-filter>
+        </activity>
+
+        <activity
             android:name="com.best.friends.splash.SplashActivity"
             android:exported="true">
             <intent-filter>
