+++ build.gradle
@@ -10,8 +10,8 @@ buildscript {
     dependencies {
         classpath 'org.jlleitschuh.gradle:ktlint-gradle:10.3.0'
         classpath 'com.google.dagger:hilt-android-gradle-plugin:2.42'
-        classpath 'com.google.gms:google-services:4.3.10'
-        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.8.1'
+        classpath 'com.google.gms:google-services:4.3.13'
+        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.9.1'
     }
 }
 
@@ -24,4 +24,10 @@ plugins {
 
 task clean(type: Delete) {
     delete rootProject.buildDir
+}
+
+ext {
+    schemes = [
+            bestfriends  : "bestfriends"
+    ]
 }
