+++ core-deeplink/src/main/java/com/yapp/android2/deeplink/AbstractDeepLinkModule.kt
@@ -3,4 +3,4 @@ package com.yapp.android2.deeplink
 import com.airbnb.deeplinkdispatch.DeepLinkModule
 
 @DeepLinkModule
-abstract class DeepLinkModule
+abstract class AbstractDeepLinkModule
