+++ presentation/home/src/main/java/com/best/friends/home/MainDeepLinkModule.kt
@@ -1,0 +1,7 @@
+package com.best.friends.home
+
+import com.airbnb.deeplinkdispatch.DeepLinkModule
+import com.yapp.android2.deeplink.AbstractDeepLinkModule
+
+@DeepLinkModule
+class MainDeepLinkModule : AbstractDeepLinkModule()
