+++ presentation/login/src/main/java/com/best/friends/login/LoginDeepLinkModule.kt
@@ -1,0 +1,7 @@
+package com.best.friends.login
+
+import com.airbnb.deeplinkdispatch.DeepLinkModule
+import com.yapp.android2.deeplink.AbstractDeepLinkModule
+
+@DeepLinkModule
+class LoginDeepLinkModule : AbstractDeepLinkModule()
