+++ presentation/notification/src/main/java/com/best/friends/notification/NotificationDeepLinkModule.kt
@@ -1,0 +1,7 @@
+package com.best.friends.notification
+
+import com.airbnb.deeplinkdispatch.DeepLinkModule
+import com.yapp.android2.deeplink.AbstractDeepLinkModule
+
+@DeepLinkModule
+class NotificationDeepLinkModule : AbstractDeepLinkModule()
